refactor: api, route, style 상수 분리
