+++ src/constants/style.ts
@@ -1,0 +1,9 @@
+/** 하단 네비게이션 사이즈 */
+export const bottomNavSize = '60px';
+
+export const HeadingStyle = {
+   main: {
+      HeadingStyle: 'mt-[40px] mb-[3px] text-center',
+      subHeadingStyle: 'mb-[65px] text-center font-normal text-[11px]',
+   },
+};
