+++ src/pages/mypage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { API_PATH } from 'constant';
+import { API_PATH } from 'constants/api';
 import { useEffectOnce } from 'hooks/useEffectOnce';
 import Box from 'components/ui/box';
 import Button from 'components/ui/button';
@@ -36,12 +36,13 @@ export default function MyPage() {
    useEffectOnce(() => {
       fetchMyInfo();
       setLayout({
-         topHeader: true,
          title: '마이페이지',
          backButton: true,
          isMain: false,
-         background: true,
          fullscreen: false,
+         headingStyle: '',
+         subHeadingStyle: '',
+         margin: '',
          rounded: true,
       });
    });
