+++ src/pages/notice/[id].tsx
@@ -1,6 +1,6 @@
 import Carousel from 'components/common/carousel';
 import PostBox, { FileBox } from 'components/ui/box/PostBox';
-import { API_PATH } from 'constant';
+import { API_PATH } from 'constants/api';
 import React from 'react';
 import Collapse from 'components/ui/collapse';
 import { useFetchPost } from 'hooks/useFetchPost';
